refactor: 도메인별 디렉토리 구조를 사용하도록 User 엔티티 클래스의 위치 변경
